Refactor remove method in RemoveDependentService to include error handling and validation

+++ src/modules/dependents/services/remove.dependent.service.ts
@@ -1,10 +1,22 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import DependentRepository from '../typeorm/repositories/DependentRepository';
 
 @Injectable()
 export class RemoveDependentService {
   constructor(private readonly dependentRepository: DependentRepository) {}
-  remove(id: string) {
+  async remove(id: string) {
+    const dependent = await this.dependentRepository.findOne(id);
+
+    if (!dependent) {
+      throw new HttpException('Dependente não encontrado', 404);
+    }
+
+    console.log('dependent: ', dependent);
+
+    if (dependent.patients[0].appointments.length > 0) {
+      throw new HttpException('Dependente possui atendimentos', 400);
+    }
+
     return this.dependentRepository.delete(id);
   }
 }
